+++ src/main/java/com/project/app/entities/Hobby.java
@@ -14,12 +14,13 @@ import java.util.Set;
 public class Hobby {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private Long idHobby;
+    private int idHobby;
     private String hobbyName;
     @Nullable
     private String description;
 
     @ManyToMany(mappedBy = "hobbySet",fetch = FetchType.EAGER)
     private Set<User> userSet= new HashSet<>();
+    
 
 }
