+++ src/main/java/com/project/app/repositories/EventRelationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRelationRepository extends JpaRepository<EventRelation,Long> {
+public interface EventRelationRepository extends JpaRepository<EventRelation,Integer> {
 }
