+++ src/main/java/com/project/app/repositories/ForumEventRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ForumEventRepository extends JpaRepository<ForumEvent,Long> {
+public interface ForumEventRepository extends JpaRepository<ForumEvent,Integer> {
 }
