+++ src/main/java/com/project/app/repositories/RelationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RelationRepository extends JpaRepository<Relation,Long> {
+public interface RelationRepository extends JpaRepository<Relation,Integer> {
 }
